refactor: вынесены тесты для hello endpoint
